feat(api): move get-answerable-elements to utils/
and delete list-as-csv function
